+++ packages/simplytics/lib/src/crash_reporting/information_properties/error_tag.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/foundation.dart';
+import 'package:simplytics/src/crash_reporting/information_properties/error_information_property.dart';
+
+/// Allows you to tag an error message when reporting it to the error monitoring system.
+///
+/// See also:
+///
+///  * [SimplyticsCrashlogInterface.recordError] and [SimplyticsCrashlogInterface.recordFatalError],
+/// which uses in the information parameter.
+class SimplyticsErrorTag extends SimplyticsErrorInformationProperty {
+  /// Creates a tag with the "name" and the "value" for reporting to the error monitoring system.
+  SimplyticsErrorTag(String name, Object value)
+      : super(name, value, level: DiagnosticLevel.hint);
+
+  @override
+  String get name => '#${super.name}';
+}