+++ packages/simplytics/lib/src/crash_reporting/services/debug_crashlog_service.dart
@@ -28,11 +28,20 @@ class SimplyticsDebugCrashlogService extends SimplyticsCrashlogInterface {
   }
 
   @override
-  Future<void> recordError(exception, StackTrace? stackTrace,
-      {reason, bool fatal = false}) async {
+  Future<void> recordError(
+    dynamic exception,
+    StackTrace? stackTrace, {
+    dynamic reason,
+    Iterable<Object> information = const [],
+    bool fatal = false,
+  }) async {
     if (_enabled) {
+      final String info = information.isEmpty
+          ? ''
+          : ('\n${StringBuffer()..writeAll(information, '\n')}');
+
       developer.log(
-        '• ${fatal ? 'Fatal ' : ''}Error${(reason != null) ? ': $reason' : ''}',
+        '• ${fatal ? 'Fatal ' : ''}Error${(reason != null) ? ': $reason' : ''}$info',
         error: exception,
         stackTrace: stackTrace,
         name: _serviceName,