+++ packages/simplytics_firebase/lib/src/firebase_crashlog_service.dart
@@ -21,10 +21,20 @@ class SimplyticsFirebaseCrashlogService extends SimplyticsCrashlogInterface {
   }
 
   @override
-  Future<void> recordError(exception, StackTrace? stackTrace,
-      {reason, bool fatal = false}) {
-    return crashlytics.recordError(exception, stackTrace,
-        reason: reason, fatal: fatal);
+  Future<void> recordError(
+    dynamic exception,
+    StackTrace? stackTrace, {
+    dynamic reason,
+    Iterable<Object> information = const [],
+    bool fatal = false,
+  }) {
+    return crashlytics.recordError(
+      exception,
+      stackTrace,
+      reason: reason,
+      information: information,
+      fatal: fatal,
+    );
   }
 
   @override