+++ packages/simplytics_sentry/example/android/app/src/main/kotlin/com/example/simplytics_firebase_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.simplytics_sentry_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()