+++ packages/simplytics/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.0
+
+* Added the ability to attach additional information to error reports, such as tags, custom properties, etc.
+
 ## 2.1.0
 
 * Fix parameter passing to FlutterError error handlers inside runAppGuarded.