+++ packages/simplytics_firebase/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.0
+
+* Added the ability to attach additional information to error reports, such as tags, custom properties, etc.
+
 ## 2.0.0
 
 * Improved compatibility with Flutter 3.10.