+++ packages/simplytics_sentry/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.0.0
+
+* First public release
+
 ## 1.0.0
 
 * Initial release